core: move CachedBlockMaxSpeedEnvBuilder to stdcm
