+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/VisitedNodesTests.kt
@@ -5,7 +5,6 @@ import fr.sncf.osrd.stdcm.graph.StopTimeData
 import fr.sncf.osrd.stdcm.graph.TimeData
 import fr.sncf.osrd.stdcm.graph.visited_node_tracking.VisitedNodes
 import fr.sncf.osrd.stdcm.infra_exploration.EdgeIdentifier
-import fr.sncf.osrd.utils.CachedBlockMaxSpeedEnvBuilder
 import fr.sncf.osrd.utils.DummyInfra
 import fr.sncf.osrd.utils.units.meters
 import kotlin.test.assertEquals
